backend/app/requirements.txt
fastapi
uvicorn
python-binance
python-dotenv
pydantic
requests
python-multipart
aiofiles
pandas
binance_client